+++ src/main/java/com/foxminded/university/models/Group.java
@@ -19,7 +19,7 @@ public class Group {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "groupname")
     private String groupName;
@@ -34,16 +34,16 @@ public class Group {
 
     }
 
-    public Group(int id, String groupName) {
+    public Group(long id, String groupName) {
         this.id = id;
         this.groupName = groupName;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
