+++ src/main/java/com/foxminded/university/models/Student.java
@@ -19,7 +19,7 @@ public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
 
     @Column(name = "firstname")
     private String firstName;
@@ -35,18 +35,18 @@ public class Student {
 
     }
 
-    public Student(int id, String firstName, String lastName, Group group) {
+    public Student(long id, String firstName, String lastName, Group group) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
         this.group = group;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
